refactor draw functions
